+++ Sources/MapboxMaps/Foundation/MapboxMap.swift
@@ -1053,6 +1053,13 @@ extension MapboxMap {
 
 }
 
+// MARK: - Map Recorder
+extension MapboxMap {
+    internal func makeRecorder() -> MapRecorder {
+        MapRecorder(mapView: __map)
+    }
+}
+
 // MARK: - Testing only!
 
 extension MapboxMap {